+++ packages/dev/cli/create.ts
@@ -1,12 +1,13 @@
 /* eslint-disable no-prototype-builtins */
 import { execSync } from 'child_process';
-import { blueBright, green, red } from 'colorette';
+import { blueBright, green, red, white } from 'colorette';
 import { cpSync } from 'fs';
 import pkg from 'fs-extra';
 import ora from 'ora';
 import { dirname, resolve } from 'path';
 import { fileURLToPath } from 'url';
 
+import { getPkgVersion } from './getPkgVersion.js';
 import type { PWAFeatures } from './run.js';
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -34,8 +35,6 @@ async function integrateServiceWorker(
 ) {
   const templateDir = resolve(__dirname, 'templates');
 
-  console.log('Integrating Service Worker...'); // todo: ora spinners for each step
-
   if (precache) {
     // if (workbox) { return; }
     const workerDir = resolve(projectDir, dir, `entry.worker.${lang}`);
@@ -43,7 +42,7 @@ async function integrateServiceWorker(
     if (pkg.pathExistsSync(workerDir)) {
       console.log(red('Service worker already exists'));
     } else {
-      const workerContent = pkg.readFileSync(resolve(templateDir, 'app', `precache.worker.${lang}`), 'utf-8');
+      const workerContent = await pkg.readFile(resolve(templateDir, 'app', `precache.worker.${lang}`), 'utf-8');
 
       await pkg.writeFile(workerDir, workerContent, 'utf-8');
     }
@@ -55,7 +54,7 @@ async function integrateServiceWorker(
     if (pkg.pathExistsSync(workerDir)) {
       console.log(red('Service worker already exists'));
     } else {
-      const workerContent = pkg.readFileSync(resolve(templateDir, 'app', `entry.worker.${lang}`), 'utf-8');
+      const workerContent = await pkg.readFile(resolve(templateDir, 'app', `entry.worker.${lang}`), 'utf-8');
 
       await pkg.writeFile(workerDir, workerContent, 'utf-8');
     }
@@ -65,8 +64,6 @@ async function integrateServiceWorker(
 async function integrateManifest(projectDir: string, lang: 'ts' | 'js' = 'ts', dir: string = 'app') {
   const templateDir = resolve(__dirname, 'templates');
   const manifestDir = resolve(projectDir, dir, `routes/manifest[.]webmanifest.${lang}`);
-
-  console.log('Integrating Web Manifest...');
 
   if (pkg.pathExistsSync(manifestDir)) {
     return;
@@ -85,6 +82,7 @@ async function integrateIcons(projectDir: string) {
 // temporary
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 async function integratePush(projectDir: string, lang: 'ts' | 'js' = 'ts', dir: string = 'app') {
+  console.log('Push API is still coming to v3.0 later...');
   return null;
 }
 
@@ -141,24 +139,40 @@ export async function createPWA(
     switch (feature) {
       case 'sw':
         try {
+          const spinnerWorker = ora({
+            text: white(`Integrating Service Worker...\n`),
+            spinner: 'dots',
+          }).start();
+
           await integrateServiceWorker(projectDir, precache, workbox, lang, dir);
+
+          spinnerWorker.succeed(`Successfully integrated Service Worker!`);
+          spinnerWorker.clear();
         } catch (err) {
           console.log(typeof err === 'string' ? red(err) : err);
         }
         break;
       case 'manifest':
         try {
+          const spinnerManifest = ora({
+            text: white(`Integrating Web Manifest...\n`),
+            spinner: 'dots',
+          }).start();
+
           await integrateManifest(projectDir, lang, dir);
+
+          spinnerManifest.succeed(`Successfully integrated Web Manifest!`);
+          spinnerManifest.clear();
         } catch (err) {
           console.log(typeof err === 'string' ? red(err) : err);
         }
         break;
       case 'icons':
-        integrateIcons(projectDir);
+        await integrateIcons(projectDir);
         break;
       case 'push':
         push = true;
-        integratePush(projectDir, lang, dir);
+        await integratePush(projectDir, lang, dir);
         break;
       case 'utils':
         utils = true;
@@ -188,28 +202,32 @@ export async function createPWA(
     json.scripts = {};
   }
 
-  json.dependencies['@remix-pwa/worker-runtime'] = '^1.0.1';
+  json.dependencies['@remix-pwa/worker-runtime'] = `^${
+    _isTest ? '' : await getPkgVersion('@remix-pwa/worker-runtime')
+  }`;
   json.dependencies.dotenv = '^16.0.3';
 
-  json.devDependencies['@remix-pwa/dev'] = '^2.0.0';
+  json.devDependencies['@remix-pwa/dev'] = `${_isTest ? '' : await getPkgVersion('@remix-pwa/dev')}`;
+  json.devDependencies['remix-pwa'] = `^${_isTest ? '' : await getPkgVersion('remix-pwa')}`;
   json.devDependencies['npm-run-all'] = '^4.1.5';
+  json.devDependencies.glob = `^${_isTest ? '' : await getPkgVersion('glob')}`;
 
   if (features.includes('sw')) {
-    json.dependencies['@remix-pwa/cache'] = '^2.0.0';
-    json.dependencies['@remix-pwa/sw'] = '^2.0.0';
-    json.dependencies['@remix-pwa/strategy'] = '^2.0.0';
+    json.dependencies['@remix-pwa/cache'] = `^${_isTest ? '' : await getPkgVersion('@remix-pwa/cache')}`;
+    json.dependencies['@remix-pwa/sw'] = `^${_isTest ? '' : await getPkgVersion('@remix-pwa/sw')}`;
+    json.dependencies['@remix-pwa/strategy'] = `^${_isTest ? '' : await getPkgVersion('@remix-pwa/strategy')}`;
   }
 
   if (push) {
-    json.dependencies['@remix-pwa/push'] = '^2.0.0';
+    json.dependencies['@remix-pwa/push'] = `^${_isTest ? '' : await getPkgVersion('@remix-pwa/push')}`;
   }
 
   if (utils) {
-    json.dependencies['@remix-pwa/client'] = '^2.0.0';
+    json.dependencies['@remix-pwa/client'] = `^${_isTest ? '' : await getPkgVersion('@remix-pwa/client')}`;
   }
 
   if (sync) {
-    json.dependencies['@remix-pwa/sync'] = '^2.0.0';
+    json.dependencies['@remix-pwa/sync'] = `^${_isTest ? '' : await getPkgVersion('@remix-pwa/sync')}`;
   }
 
   json.scripts.build = 'run-s build:*';
@@ -224,7 +242,7 @@ export async function createPWA(
 
   if (install) {
     const spinner = ora({
-      text: blueBright(`Running ${packageManager} install...`),
+      text: blueBright(`Running ${packageManager} install...\n`),
       spinner: 'dots',
     }).start();
 
@@ -234,7 +252,7 @@ export async function createPWA(
         stdio: 'inherit',
       });
 
-      spinner.succeed(`Successfully ran ${packageManager} install!`);
+      spinner.succeed(`Successfully ran ${packageManager} install!\n`);
       spinner.clear();
     } else {
       spinner.succeed(`Successfully installed dependencies!`).clear();