+++ packages/worker-runtime/CHANGELOG.md
@@ -1,3 +1,17 @@
+## @remix-pwa/worker-runtime 2.0.4 (2023-09-21)
+
+
+### Bug Fixes
+
+* **release:** force a new release 93dcbe7
+
+## @remix-pwa/worker-runtime 2.0.2 (2023-09-21)
+
+
+### Bug Fixes
+
+* **typescript:** use new types from remix@2 db36fce
+
 ## @remix-pwa/worker-runtime 2.0.1 (2023-09-15)